refactor(queue): move vite playground to workspace root
