+++ package.json
@@ -14,6 +14,7 @@
     "typecheck": "pnpm --dir docs typecheck && pnpm -r --filter './packages/*' --if-present run typecheck",
     "test": "pnpm -r --filter './packages/*' --if-present run test",
     "build": "pnpm -r --filter './packages/*' --if-present run build",
+    "playground:vite:build": "pnpm --dir playground/vite build",
     "docs:dev": "pnpm --dir docs dev",
     "docs:build": "pnpm --dir docs build"
   },
