--- packages/queue/playground/vite/noop.ts
@@ -1,1 +1,0 @@
-export {}
